fix: remove autofocus from text editor as create conversation form opens. - Adds new prop autofocus on text editor - rename ConversationTextEditor to TextEditor
