+++ i18n/en.json
@@ -589,13 +589,14 @@
   "conversation.sidebar.previousConvo": "Previous conversations",
   "conversation.sidebar.noPreviousConvo": "No previous conversations",
   "conversation.sidebar.notAvailable": "Not available",
-  "editor.placeholder": "Shift + Enter to add a new line",
+  "editor.newLine": "Shift + Enter to add a new line. ",
+  "editor.send": " Cmd + Enter to send. ",
+  "editor.cmdK": "Cmd + K to open command bar. ",
   "ai.apiKeyNotSet": "{provider} API Key is not set. Please ask administrator to set it up",
   "ai.enterOpenAIAPIKey": "Enter OpenAI API Key",
   "ai.apiKey.description": "{provider} API Key is not set or invalid. Please enter a valid API key to use AI features.",
   "replyBox.reply": "Reply",
   "replyBox.emailAddresess": "Email addresses separated by comma",
-  "replyBox.editor.placeholder": "Shift + Enter to add a new line. Cmd + Enter to send. Cmd + K to open command bar.",
   "replyBox.invalidEmailsIn": "Invalid email(s) in",
   "replyBox.correctEmailErrors": "Please correct the email errors before sending.",
   "contact.blockConfirm": "Are you sure you want to block this contact? They will no longer be able to interact with you.",
